PPLM

Documentation for PPLM.

PPLM.GPT2Tokenizer

PPLM.build_bow_ohe

PPLM.data_preprocess

PPLM.get_artifact

PPLM.get_bow_indices

PPLM.get_mask

PPLM.get_registered_file

PPLM.load_cached_data

PPLM.load_data

PPLM.load_data_from_csv

PPLM.pad_seq

PPLM.register_custom_file

PPLM.save_classifier_head

PPLM.top_k_logits

PPLM.truncate_

PPLM.GPT2Tokenizer — Method

Example: for vector of texts -> map(x->encode(tokenizer, x), textvector) or tokenizer.(textvector)

PPLM.build_bow_ohe — Method

build_bow_ohe(bow_indices, tokenizer)

Build and return a list of one_hot_matrix for each Bag Of Words list from indices. Each item of the list is of dimension (num_of_BoW_words, tokenizer.vocab_size).

Note: While building the OHE of word indices, it only keeps those words, which have length 1 after tokenization and discard the rest.

PPLM.data_preprocess — Function

data_preprocess(data_x, data_y, classification_type::String="Binary", num_classes::Integer=2; args=nothing)

Function to preprocess data_x and data_y along with creating mask for the data_x.

Preprocessing for data_x consist of padding with pad token (expected to be provided as args.pad_token).

Preprocessing for data_y consist of creating onehotbach for data_y (if classification_type is not "Binary"), for 1:num_classes else reshape the data as (1, length(data_y))

Returns data_x, data_y, mask after pre-processing.

PPLM.get_artifact — Method

get_artifact(name)

Utility function to download/install the artifact in case not already installed.

PPLM.get_bow_indices — Method

get_bow_indices(bow_key_or_path_list::Vector{String}, tokenizer)

Returns a list of list of indices of words from each Bag of word in the bow_key_or_path_list, after tokenization. The functions looks for provided BoW key in the registered artifacts Artifacts.toml file. In case not present there, function expects that bow_key is provided as the complete path to the file the URL to download .txt file.

Example

get_bow_indices(["legal", "military"])

PPLM.get_mask — Method

get_mask(seq::AbstractMatrix{T}, pad_token::Integer=0, embed_size::Integer=768)

Function to create mask for sequences against padding, so as to inform the model, that some part of sequenece is padded and hence to be ignored.

PPLM.get_registered_file — Method

get_registered_file(name)

Fetch registered file path from Artifacts.toml, based on the artifact name.

PPLM.load_cached_data — Method

load_cached_data(discrim::Union{DiscriminatorV1, DiscriminatorV2}, data_x, data_y, tokenizer::PretrainedTokenizer; truncate::Bool=false, max_length::Integer=256, shuffle::Bool=false, batchsize::Int=4, drop_last::Bool=false, classification_type="Binary", num_classes=2, args=nothing)

Returns a DataLoader with (x, y) which can directly be feeded into classifier layer for training.

The function first loads the data using load_data function with batchsize=1, then passes each batch to the transformer model of discrim after data preprocessing, and then the average representation of the hidden_states are stored in a vector, which are then further loaded into a DataLoader, ready to use for classification training.

Note: This functions saves time by cacheing the average representation of hidden states beforehand, avoiding passing the data through model in each epoch of training. This can be done as the model itself is non-trainable while training discriminator classifier head.

PPLM.load_data — Method

load_data(data_x, data_y, tokenizer::PretrainedTokenizer;  batchsize::Integer=8, truncate::Bool=false, max_length::Integer=256, shuffle::Bool=false, drop_last::Bool=false, add_eos_start::Bool=true)

Returns DataLoader for the data_x and data_y after processing the datax, with batchsize=batchsize. The processing consist of tokenization of datax and further truncation to max_len if truncate is set to be true.

If add_eos_start is set to true, add EOS token of tokenizer to the start.

PPLM.load_data_from_csv — Method

load_data_from_csv(path_to_csv; text_col="text", label_col="label", delim=',', header=1)

Load the data from a csv file based on the specified text_col column for text and label_col for target label. Returns vectors for text and label.

PPLM.pad_seq — Method

pad_seq(batch::AbstractVector{T}, pad_token::Integer=0)

Function to add pad tokens in shorter sequence, to make the length of each sequence equal to the max_length ( calculated as max(map(length, batch))) in the batch. Pad token defaults to 0.

PPLM.register_custom_file — Method

register_custom_file(artifact_name, file_name, path)

Function to register custom file under artifact_name in Artifacts.toml. path expects path of the directory where the file file_name is stored. Stores the complete path to the file as Artifact URL.

Example

register_custom_file('custom', 'xyz.txt','./folder/folder/')

Note: In case this gives permission denied error, change the Artifacts.toml file permissions using chmod(path_to_file_in_julia_installation , 0o764)or similar.

PPLM.save_classifier_head — Method

PPLM.top_k_logits — Method

top_k_logits(logits::AbstractArray, k; prob = false)

Masks everything but the k top entries as -infinity (1e10). Incase of probs=true, everthing except top-k probabilities are masked to 0.0. logits is expected to be a vector.

PPLM.truncate_ — Method

truncate_(x, max_length::Integer)

Truncate the data to minimum of max_length and length of x.